requirements.txt
altgraph==0.17.2
asgiref==3.5.1
auto-py-to-exe==2.19.0
backports.zoneinfo==0.2.1
bottle==0.12.19
bottle-websocket==0.2.9
Django==4.0.4
djangorestframework==3.13.1
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
